// src/configs/global.enum.ts
export enum Post {
  POST_CALLER = 'POST_CALLER',
  POST_DEFAULT = 'POST_DEFAULT',
  POST_STARTED = 'POST_STARTED',
  POST_FAILED = 'POST_FAILED',
  POST_SUCCESS = 'POST_SUCCESS',
}

export enum GetCategories {
  GET_CATEGORIES_CALLER = 'GET_CATEGORIES_CALLER',
  GET_CATEGORIES_DEFAULT = 'GET_CATEGORIES_DEFAULT',
  GET_CATEGORIES_STARTED = 'GET_CATEGORIES_STARTED',
  GET_CATEGORIES_FAILED = 'GET_CATEGORIES_FAILED',
  GET_CATEGORIES_SUCCESS = 'GET_CATEGORIES_SUCCESS',
}
// ...
export enum searchLocation {
  SEARCH_LOCATION_CALLER = 'DO-SEARCH-LOCATION ',
  SEARCH_LOCATION_DEFAULT = 'SEARCH-LOCATION-NOT-STARTED',
  SEARCH_LOCATION_STARTED = 'SEARCH-LOCATION-STARTED',
  SEARCH_LOCATION_FAILED = 'SEARCH-LOCATION-FAILED',
  SEARCH_LOCATION_SUCCESS = 'SEARCH-LOCATION-SUCCESS',
}

export enum setLocation {
  SET_LOCATION_CALLER = 'DO-SET_LOCATION ',
  SET_LOCATION_DEFAULT = 'SET_LOCATION-NOT-STARTED',
  SET_LOCATION_STARTED = 'SET_LOCATION-STARTED',
  SET_LOCATION_FAILED = 'SET_LOCATION-FAILED',
  SET_LOCATION_SUCCESS = 'SET_LOCATION-SUCCESS',
}
// ...
export enum inputActionType {
  //SAGA CALLERS
  SET_LOGIN_PHONE_NUMBER_CALLER = 'SET-LOGIN-PHONE-NUMBER',
  SET_LOGIN_PASSWORD_CALLER = 'SET-LOGIN-PASSWORD',
  SET_FIRST_NAME_CALLER = 'SET-FIRST-NAME',
  SET_LAST_NAME_CALLER = 'SET-LAST-NAME',
  SET_SIGNUP_PHONE_NUMBER_CALLER = 'SET-SIGNUP-PHONE-NUMBER',
  SET_SIGNUP_PASSWORD_CALLER = 'SET-SIGNUP-PASSWORD',
  SET_GENDER_CALLER = 'SET-GENDER',
  SET_RESSET_PASSWORD_PHONE_NUMBER_CALLER = 'SET-RESSET-PASSWORD-PHONE-NUMBER',
  SET_VERIFICATION_PIN_CALLER = 'SET-VERIFICATION-PIN',
  SET_TIMER_ELAPSE_CALLER = 'SET-TIMER-ELAPSE',
  CHANGE_COUNTDOWN_COMPONENT_ID_CALLER = 'CHANGE-COUTDOWN-COMPOENT-ID',
  SET_NEW_PASSWORD_CALLER = 'SET-NEW-PASSWORD',
  SET_CONFIRM_NEW_PASSWORD_CALLER = 'SET-CONFIRM-NEW-PASSWORD',
  SET_PARCEL_LOCATION_CALLER = 'DO-SET-PARCEL-LOCATION',
  SET_PARCEL_DESITINATION_CALLER = 'DO-SET-PARCEL-DESTINATION',
  SET_LOCATION_ACTIVE_CALLER = 'DO-SET-LOCATION-ACTIVE',
  SET_VARIABLES_CALLER = 'DO-SET-VARIABLES',
  SET_RESET_INPUTS_CALLER = 'DO-RESET-INPUTS',
  SET_DURATION_AND_DISTANCE_CALLER = 'DO-SET-DURATION-AND-DISTANCE',
  SET_ENABLE_ADD_BUTTON_CALLER = 'DO-ENABLE-ADD-BUTTON',
  SET_ADD_VALID_CARD_CALLER = 'DO-ADD-VALID-CARD',
  SET_REMOVE_VALID_CARD_CALLER = 'DO-REMOVE-VALID-CARD',
  SET_CHANGE_PAYMENT_METHOD_CALLER = 'DO-CHANGE-PAYMENT-METHOD',
  SET_TRACK_ACTIVE_DELIVERY_CALLER = 'DO-SET-DELIVERY-ON-PROGRESS',
  SET_USER_CURRENT_LOCATION_CALLER = 'DO-SET-CURRENT-LOCATION',
  SET_LOCATION_CALLER = 'DO-SET-FETCHED-LOCATION',
  SET_PROFILE_PICTURE_CALLER = 'DO-SET-USER-PROFILE-PICTURE',
  SET_CARD_EMAIL_CALLER = 'DO_SET_CARD_EMAIL_CALLER',

  //ACTIONS
  SET_LOGIN_PHONE_NUMBER = 'SET-LOGIN-PHONE-NUMBER',
  SET_LOGIN_PASSWORD = 'SET-LOGIN-PASSWORD',
  SET_FIRST_NAME = 'SET-FIRST-NAME',
  SET_LAST_NAME = 'SET-LAST-NAME',
  SET_SIGNUP_PHONE_NUMBER = 'SET-SIGNUP-PHONE-NUMBER',
  SET_SIGNUP_PASSWORD = 'SET-SIGNUP-PASSWORD',
  SET_GENDER = 'SET-GENDER',
  SET_RESSET_PASSWORD_PHONE_NUMBER = 'SET-RESSET-PASSWORD-PHONE-NUMBER',
  SET_VERIFICATION_PIN = 'SET-VERIFICATION-PIN',
  SET_TIMER_ELAPSE = 'SET-TIMER-ELAPSE',
  CHANGE_COUNTDOWN_COMPONENT_ID = 'CHANGE-COUTDOWN-COMPOENT-ID',
  SET_NEW_PASSWORD = 'SET-NEW-PASSWORD',
  SET_CONFIRM_NEW_PASSWORD = 'SET-CONFIRM-NEW-PASSWORD',
  SET_PARCEL_LOCATION = 'SET-PARCEL-LOCATION',
  SET_PARCEL_DESITINATION = 'SET-PARCEL-DESTINATION',
  SET_LOCATION_ACTIVE = 'SET-LOCATION-ACTIVE',
  SET_VARIABLES = 'SET-VARIABLES',
  SET_RESET_INPUTS = 'RESET-INPUTS',
  SET_DURATION_AND_DISTANCE = 'SET-DURATION-AND-DISTANCE',
  SET_ENABLE_ADD_BUTTON = 'ENABLE-ADD-BUTTON',
  SET_ADD_VALID_CARD = 'ADD-VALID-CARD',
  SET_REMOVE_VALID_CARD = 'REMOVE-VALID-CARD',
  SET_CHANGE_PAYMENT_METHOD = 'CHANGE-PAYMENT-METHOD',
  SET_TRACK_ACTIVE_DELIVERY = 'SET-DELIVERY-ON-PROGRESS',
  SET_USER_CURRENT_LOCATION = 'SET-CURRENT-LOCATION',
  SET_PROFILE_PICTURE = 'SET-USER-PROFILE-PICTURE',
  SET_CARD_EMAIL = 'SET_CARD_EMAIL',
}